+++ src/structured_data/match.py
@@ -6,6 +6,7 @@ from ._match_failure import MatchFailure
 from ._not_in import not_in
 from ._patterns import DISCARD
 from ._patterns import AttrPattern
+from ._patterns import Bind
 from ._patterns import DictPattern
 from ._patterns import Pattern
 
@@ -71,11 +72,18 @@ def _match_iteration(destructurers, match_dict, target, value):
 
 def _match(target, value, destructurers):
     match_dict = MatchDict()
+    extra_bindings = ()
+    if isinstance(target, Bind):
+        extra_bindings = target.bindings
+        target = target.structure
     to_process = [(target, value)]
     while to_process:
         to_process.extend(
             _match_iteration(destructurers, match_dict, *to_process.pop())
         )
+    for (name, bind_value) in extra_bindings:
+        not_in(match_dict, name)
+        match_dict[name] = bind_value
     return match_dict